refactor(frontend): split deployments stepper context
